Add Collision system
GameObject has a Collider member. Interactions between Colliders are
processed in the first part of the loop (just before the logic part).
Colliders have two bounding box types - Point and Circle. When an
interaction is detected, a Collision object is allocated to a central
buffer owned by Main Scene. Each game loop iteration, all Collisions
are cleard. If a Collision occurs to a Projectile, it is destroyed.